services/codenames/requirements.txt
# Core dependencies
nikitas-agents>=0.1.0
fastapi>=0.115.0
uvicorn>=0.32.0
websockets>=13.0
pydantic>=2.0.0
python-dotenv>=1.0.0

